feat: comment edit function added

+++ find_thanku/src/component/comment/index.js
@@ -4,10 +4,15 @@ import * as S from "./styles";
 import { STATIC_URL } from "../../constant";
 
 const Comment = ({comment}) => {
+  const url = window.location.pathname;
   const Delete = async () => {
-    //작성자 == 현재 접속자(sohee(temp))
-    if(comment.username === "sohee(temp)"){
-      const url = window.location.pathname;
+    var bool = window.confirm('정말 삭제하시겠습니까?');
+    //삭제 안함
+    if(!bool){
+      return; 
+    }
+    //삭제함 && 작성자 == 현재 접속자의 username(sohee(temp))
+    if(bool && comment.username === "sohee(temp)"){
       const request = await fetch("http://localhost:4000" + url + "/comment/" + comment._id, {
         method: "DELETE",
       });
@@ -16,11 +21,32 @@ const Comment = ({comment}) => {
         return;
       }
       await request.json();      
-      alert('댓글 삭제'); window.location.reload(true); //새로고침
+      alert('댓글 삭제'); 
+      window.location.reload(true); //새로고침
     }
-    //작성자가 아님->삭제X
+    //작성자!= 접속자 -> 삭제 안함
     else{
       alert('삭제 권한 없음');
+    }
+  }
+ 
+  const Edit = async () => {
+    //작성자 == 현재 접속자의 username(sohee(temp))
+    if(comment.username === "sohee(temp)"){
+      var input = prompt('새로운 내용을 입력하세요',comment.content);
+      const request = await fetch("http://localhost:4000" + url + "/comment/" + comment._id +"/content/"+input, {
+        method: "PATCH",
+      });
+      if (!request.ok) {
+        alert("댓글 수정 실패");
+        return;
+      }
+      await request.json();
+      window.location.reload(true); //새로고침
+    }
+    //작성자!= 접속자 -> 수정 안함
+    else{
+      alert('수정 권한 없음');
     }
   }
   
@@ -33,10 +59,9 @@ const Comment = ({comment}) => {
         <S.Writer>{comment.username} </S.Writer>
       </S.WriterContainer>
       <S.ContentContainer>
-        <S.Content>
-          {comment.content}
-          {<img src={STATIC_URL.DELETE} alt="delete" width='20px' align='right' onClick={Delete}/>}
-        </S.Content>
+        <S.Content>{comment.content}</S.Content>
+        <S.CommentIcon>{<img src={STATIC_URL.EDIT} alt="edit" align='right' onClick={Edit}/>}</S.CommentIcon>
+        <S.CommentIcon>{<img src={STATIC_URL.DELETE} alt="delete" align='right' onClick={Delete}/>}</S.CommentIcon>
       </S.ContentContainer>
     </S.Comment>
   );
